# Performance enhancements for has(), setService(), setFactory(), setShared(), setAlias(), setInvokableClass() and configure()