--- a/src/AbstractPluginManager.php
+++ b/src/AbstractPluginManager.php
@@ -121,6 +121,15 @@ abstract class AbstractPluginManager extends ServiceManager implements PluginMan
     }
 
     /**
+     * Override setService for additional plugin validation.
+     */
+    public function setService($name, $service)
+    {
+        $this->validate($service);
+        parent::setService($name, $service);
+    }
+
+    /**
      * {@inheritDoc}
      *
      * @param string $name Service name of plugin to retrieve.