--- a/src/ServiceManager.php
+++ b/src/ServiceManager.php
@@ -436,13 +436,10 @@ class ServiceManager implements ServiceLocatorInterface
     public function setAlias($alias, $target)
     {
         $this->validate($alias);
-
         $this->aliases[$alias] = $target;
 
-        if (isset($this->resolvedAliases[$target])) {
-            $target = $this->resolvedAliases[$target];
-        }
-        $this->resolvedAliases[$alias] = $target;
+        $this->resolvedAliases[$alias] =
+            isset($this->resolvedAliases[$target]) ? $this->resolvedAliases[$target] : $target;
         if (in_array($alias, $this->resolvedAliases)) {
             $r = array_intersect($this->resolvedAliases, [ $alias ]);
             foreach ($r as $name => $service) {