--- a/src/ServiceManager.php
+++ b/src/ServiceManager.php
@@ -345,7 +345,6 @@ class ServiceManager implements ServiceLocatorInterface
             $factories = $this->createFactoriesForInvokables($config['invokables']);
 
             if (! empty($aliases)) {
-                // @todo: This is wrong! These aliases are 'resolved' already
                 $config['aliases'] = (isset($config['aliases']))
                     ? \array_merge($config['aliases'], $aliases)
                     : $aliases;
@@ -520,6 +519,7 @@ class ServiceManager implements ServiceLocatorInterface
      *
      * @param string[]|Initializer\InitializerInterface[]|callable[] $initializers
      *
+     * @return void
      */
     private function resolveInitializer($initializer)
     {
@@ -926,13 +926,6 @@ class ServiceManager implements ServiceLocatorInterface
 
     /**
      * Instantiate abstract factories for to avoid checks during service construction.
-     *
-     * @todo: To construct to avoid is not really an optimization, it's lazyness
-     * AbstractFactories are shared services. Make sure that has() is guarded against
-     * numerical parameters. Handle AbstractFactories as any other shared services.
-     *
-     * @todo: Implement a has() and get() logic which unifies the several arrays
-     * It is a desaster to isset(blah) several times on each request.
      *
      * @param string[]|Factory\AbstractFactoryInterface[] $abstractFactories
      *