… benchmarks/HasBench.php
… benchmarks/SetNewServicesBench.php
… test/CommonServiceLocatorBehaviorsTrait.php
… test/Exception/CyclicAliasExceptionTest.php
… test/ServiceManagerTest.php