This PR is about performance of the setter APIs (or injector APIs if you like) as well as has(). Here's the current comparison report. Thanks @dantleech for faster than light fix.

	benchmark: SetNewServicesBench
	+------------------------+--------------------+------------------+--------+
	| subject                | suite:develop:mean | suite:PR221:mean | gain   |
	+------------------------+--------------------+------------------+--------+
	| benchSetService        | 1.924µs            | 0.655µs          |  2,94x |
	| benchSetFactory        | 4.183µs            | 1.226µs          |  3,41x |
	| benchSetAlias          | 11.222µs           | 1.872µs          |  5,99x |
	| benchSetAliasOverrided | 34.617µs           | 1.876µs          | 18,45x |
	+------------------------+--------------------+------------------+--------+

This is the develop branch:

	Dumped result to SetNewServices.develop.xml
	suite: 133ec8900bdebd2cfb961e09ca103052bd993f16, date: 2018-01-05, stime: 22:12:29
	+---------------------+------------------------+--------+--------+--------+-----+------------+-------------+-------------+-------------+-------------+------------+--------+--------+
	| benchmark           | subject                | groups | params | revs   | its | mem_peak   | best        | mean        | mode        | worst       | stdev      | rstdev | diff   |
	+---------------------+------------------------+--------+--------+--------+-----+------------+-------------+-------------+-------------+-------------+------------+--------+--------+
	| SetNewServicesBench | benchSetService        |        | []     | 100000 | 100 | 1,375,992b | 1.910110µs  | 1.957413µs  | 1.946003µs  | 2.050120µs  | 0.028245µs | 1.44%  | 1.00x  |
	| SetNewServicesBench | benchSetFactory        |        | []     | 100000 | 100 | 1,375,992b | 4.180240µs  | 4.240743µs  | 4.229557µs  | 4.420250µs  | 0.030047µs | 0.71%  | 2.17x  |
	| SetNewServicesBench | benchSetAlias          |        | []     | 100000 | 100 | 1,375,992b | 11.440650µs | 11.625965µs | 11.601287µs | 12.160700µs | 0.099540µs | 0.86%  | 5.94x  |
	| SetNewServicesBench | benchSetAliasOverrided |        | []     | 100000 | 100 | 1,376,000b | 35.572030µs | 35.950056µs | 35.759907µs | 37.172120µs | 0.312706µs | 0.87%  | 18.37x |
	+---------------------+------------------------+--------+--------+--------+-----+------------+-------------+-------------+-------------+-------------+------------+--------+--------+

Here comes this PR:

	Dumped result to SetNewServices.setAlias.xml
	suite: 133ec89fee8638dd6d892c639ca0556b8090f9b8, date: 2018-01-05, stime: 21:42:44
	+---------------------+------------------------+--------+--------+--------+-----+------------+------------+------------+------------+------------+------------+--------+-------+
	| benchmark           | subject                | groups | params | revs   | its | mem_peak   | best       | mean       | mode       | worst      | stdev      | rstdev | diff  |
	+---------------------+------------------------+--------+--------+--------+-----+------------+------------+------------+------------+------------+------------+--------+-------+
	| SetNewServicesBench | benchSetService        |        | []     | 100000 | 100 | 1,347,528b | 0.640030µs | 0.654438µs | 0.650130µs | 0.680040µs | 0.010230µs | 1.56%  | 1.00x |
	| SetNewServicesBench | benchSetFactory        |        | []     | 100000 | 100 | 1,347,528b | 1.200070µs | 1.222270µs | 1.217291µs | 1.280070µs | 0.013681µs | 1.12%  | 1.87x |
	| SetNewServicesBench | benchSetAlias          |        | []     | 100000 | 100 | 1,347,528b | 1.810100µs | 1.845307µs | 1.848852µs | 1.900110µs | 0.015841µs | 0.86%  | 2.82x |
	| SetNewServicesBench | benchSetAliasOverrided |        | []     | 100000 | 100 | 1,347,536b | 1.780100µs | 1.803603µs | 1.794896µs | 1.870110µs | 0.019256µs | 1.07%  | 2.76x |
	+---------------------+------------------------+--------+--------+--------+-----+------------+------------+------------+------------+------------+------------+--------+-------+